adding statistics and countdown

+++ app/src/main/java/biz/eventually/atpl/data/network/TopicDtoNetwork.kt
@@ -9,5 +9,5 @@ data class TopicDtoNetwork(
         val questions: Int = 0,
         val follow: Int?,
         val focus: Int?,
-        val mean:Int?
+        val mean:Double?
 )
